requirements.txt
python-telegram-bot[job-queue]==20.7
httpx==0.25.2
python-dotenv==1.0.0
aiohttp==3.9.1
aiosqlite==0.19.0
feedparser==6.0.11
beautifulsoup4==4.12.2
Pillow==10.2.0
